LayoutTests/platform/ios-simulator/css3/device-adapt/opera/constrain-009-expected.txt


FAIL CSS Test: @viewport constrained - width 20em assert_equals: expected 320 but got 800(stack: assert@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:2168:73
assert_equals@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:705:15
file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/device-adapt/opera/constrain-009.xhtml:69:22
step@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:1186:30
onload@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/device-adapt/opera/constrain-009.xhtml:68:14)
